requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=3.13
Pillow>=8.4
django-cors-headers>=3.2